api_bridge/requirements.txt
requests
pymysql
python-dotenv
tqdm